pyproject.toml
[tool.poetry]
name = "andreiproj"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
peewee-aio = "^2.0.0"
fastapi = "^0.115.6"
pytest = "^8.3.4"
modconfig = "^1.2.1"
pyjwt = "^2.10.1"
marshmallow = "^3.23.1"
aiosqlite = "^0.20.0"
httpx = "^0.28.1"
pytest-env = "^1.1.5"
pytest-aio = "^1.9.0"
uvicorn = "^0.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-xslv"
python_files = "*test*.py"
env = ["ENV=tests"]
